+++ src/bootloader/secondary/string.h
@@ -2,14 +2,14 @@
 
 #include <stdint.h>
 
-void memcpy(const void* srcptr, void* dstptr, uint16_t len);
+void memcpy(const void* srcptr, void* dstptr, uint32_t len);
 
-void memset(void* ptr, uint8_t value, uint16_t len);
+void memset(void* ptr, uint8_t value, uint32_t len);
 
-int16_t memcmp(const void* left, const void* right, uint16_t len);
+int32_t memcmp(const void* left, const void* right, uint32_t len);
 
-uint16_t strncpy(const char* srcptr, char* dstptr, uint16_t num);
+uint32_t strncpy(const char* srcptr, char* dstptr, uint32_t num);
 
-const char* strnchr(const char* str, char chr, uint16_t num);
+const char* strnchr(const char* str, char chr, uint32_t num);
 
-uint16_t strlen(const char* str);
+uint32_t strlen(const char* str);
